Move constants.h to src/common/
