+++ src/InOneWeekend/main.cc
@@ -10,7 +10,7 @@
 //==================================================================================================
 
 #include "camera.h"
-#include "constants.h"
+#include "common/constants.h"
 #include "hittable_list.h"
 #include "material.h"
 #include "random.h"
