+++ src/TheNextWeek/hittable.h
@@ -12,7 +12,7 @@
 //==================================================================================================
 
 #include "aabb.h"
-#include "constants.h"
+#include "common/constants.h"
 
 #include <float.h>
