+++ src/TheNextWeek/main.cc
@@ -13,8 +13,8 @@
 #include "box.h"
 #include "bvh.h"
 #include "camera.h"
+#include "common/constants.h"
 #include "constant_medium.h"
-#include "constants.h"
 #include "hittable_list.h"
 #include "material.h"
 #include "moving_sphere.h"
