--- src/TheRestOfYourLife/constants.h
@@ -1,9 +1,0 @@
-#ifndef CONSTANTSH
-#define CONSTANTSH
-
-#include <limits>
-
-const double pi = 3.1415926535897932385;
-const double infinity = std::numeric_limits<double>::infinity(); 
-
-#endif
