+++ src/TheRestOfYourLife/material.h
@@ -11,7 +11,7 @@
 // with this software. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
 //==================================================================================================
 
-#include "constants.h"
+#include "common/constants.h"
 #include "hittable.h"
 #include "onb.h"
 #include "pdf.h"
